+++ anssi/Cargo.toml
@@ -1,46 +1,32 @@
 [package]
-authors = ["Bruno Grieder <bruno.grieder@cosmian.com>"]
+authors = ["Bruno Grieder <bruno.grieder@cosmian.com>", "Théophile BRÉZOT <theophile.brezot@cosmian.com>"]
 build = "build.rs"
-description = "Crypto lib for hybrid encryption and pure crypto primitives"
+description = "Crypto lib for pure crypto primitives"
 edition = "2021"
 license = "MIT/Apache-2.0"
-name = "cosmian_crypto_base"
-version = "2.1.0"
+name = "cosmian_crypto_base_anssi"
+version = "1.0.0"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
-name = "cosmian_crypto_base"
+name = "cosmian_crypto_base_anssi"
 path = "src/lib.rs"
 
 [dependencies]
 aes = "0.7"
 aes-gcm = "0.9"
-cosmian_fpe = "0.5.1"
 curve25519-dalek = "3.2"
-# specify the js feature for the WASM target
-getrandom = { version = "0.2", features = ["js"] }
 hex = "0.4"
 hkdf = "0.12"
-itertools = "0.10"
 num-bigint = { version = "0.4", features = ["rand", "serde"] }
 num-traits = "0.2"
 rand = "0.8"
 rand_core = { version = "0.5", features = ["getrandom"] }
-rand_distr = "0.4"
 rand_hc = "0.3"
 serde = { version = "1.0", features = ["derive"] }
 sha2 = "0.10"
 thiserror = "1.0"
-tracing = "0.1"
 zeroize = "1.5.6"
-cosmian_crypto_base_anssi = { path = "anssi" }
-
-[dev-dependencies]
-retry_panic = { git = "https://github.com/Cosmian/retry_panic" }
-
-[features]
-default = []
-libsodium = []
 
 [target.'cfg(all(not(target_arch = "wasm32"), not(windows)))'.dependencies]
 cupid = "0.6"
