+++ anssi/src/symmetric_crypto/key.rs
